fix(S10): UI bugs
