--- src/components/Product/components/ProductCharacteristics/components/UI/ProductCharacteristicsRow/ProductCharacteristicsRow.module.scss
@@ -1,37 +1,0 @@
-@import "@/styles/media";
-
-.productCharacteristicsItemWrapper {
-  display: flex;
-  width: 100%;
-  flex-direction: column;
-  gap: 12px;
-  transition: width .3s;
-
-  .productCharacteristicsItem {
-    display: flex;
-    transition: color .3s;
-
-    &:hover {
-      .productCharacteristicsItemKey {
-        color: var(--secondary-c2-color-500);
-      }
-    }
-
-    .productCharacteristicsItemKey {
-      color: var(--primary-c1-color-700);
-      margin-right: 20px;
-      width: 110px;
-      transition: color .3s;
-
-      @include desktop {
-        min-width: 75px;
-        max-width: 75px;
-      }
-
-      @include productCardsContainer {
-        min-width: 110px;
-        max-width: 110px;
-      }
-    }
-  }
-}
