--- src/components/Product/components/ProductCharacteristics/components/UI/ProductCharacteristicsRow/index.ts
@@ -1,1 +1,0 @@
-export { ProductCharacteristicsRow } from "./ProductCharacteristicsRow.tsx";
