--- src/components/Product/components/ProductCharacteristics/components/UI/ProductCharacteristicsRows/ProductCharacteristicsRows.tsx
@@ -1,18 +1,0 @@
-import { FC } from "react";
-import { ProductCharacteristicsRow } from "../ProductCharacteristicsRow";
-import { ProductCharacteristicsListProps, ProductCharacteristicsOptions } from "@/common/constants";
-
-export const ProductCharacteristicsRows: FC<Omit<ProductCharacteristicsListProps, "rowIndex">> = ({
-  productCharacteristicsKeys,
-  productCharacteristicsList,
-}) => {
-  const rows = Array.from({ length: Math.ceil(productCharacteristicsKeys.length / ProductCharacteristicsOptions.COLUMNS) });
-  return rows.map((_, rowIndex) => (
-    <ProductCharacteristicsRow
-      key={rowIndex}
-      rowIndex={rowIndex}
-      productCharacteristicsKeys={productCharacteristicsKeys}
-      productCharacteristicsList={productCharacteristicsList}
-    />
-  ));
-};
