--- src/components/Product/components/ProductCharacteristics/components/UI/ProductCharacteristicsRows/index.ts
@@ -1,1 +1,0 @@
-export { ProductCharacteristicsRows } from "./ProductCharacteristicsRows.tsx";
