--- src/components/Product/components/ProductCharacteristics/components/UI/index.ts
@@ -1,2 +1,0 @@
-export { ProductCharacteristicsRows } from "./ProductCharacteristicsRows";
-export { ProductCharacteristicsRow } from "./ProductCharacteristicsRow";
