--- src/components/Product/components/ProductCharacteristics/components/index.ts
@@ -1,1 +1,0 @@
-export { ProductCharacteristicsRows } from "./UI";
