+++ src/components/Product/components/ProductNotification/ProductNotification.module.scss
@@ -1,7 +1,8 @@
 .productInfoNotificationContainer {
   display: flex;
   column-gap: 8px;
-  padding: 0 0 6px 15px;
+  padding-left: 15px;
+  margin-bottom: 6px;
 
   .productInfoNotification {
     padding: 4px 10px;
