+++ src/components/Product/components/ProductOrder/ProductOrder.module.scss
@@ -9,7 +9,8 @@
   border-radius: 12px;
   border: 1px solid var(--primary-c1-color-400);
   max-width: 650px;
-  padding: 12px 24px 12px 16px;
+  min-height: 92px;
+  padding: 20px 24px 20px 16px;
   margin-bottom: 24px;
 
   @include productCardsContainer {
@@ -25,7 +26,7 @@
     flex-direction: row;
     flex-wrap: wrap;
     justify-content: space-between;
-    align-items: stretch;
+    align-items: center;
     max-width: 100%;
     column-gap: 24px;
   }
