+++ src/components/Product/components/ProductOrder/ProductOrder.tsx
@@ -6,13 +6,7 @@ import { getAnimationVariant, getProductPrice, getProductUnitsMeasure } from "@/
 import { ProductPrice } from "@/common/types";
 import { ProductOrderNavigation } from "./components";
 import { ProductCardPrice } from "@/components/ProductsList/components";
-import {
-  AnimationDefaultDuration,
-  ProductUnitsMeasure,
-  ProductsAmountOfUnitsMeasure,
-  animationDefaultVariants,
-  notificationInfoDefaultTemplate,
-} from "@/common/constants";
+import { AnimationDefaultDuration, ProductUnitsMeasure, ProductsAmountOfUnitsMeasure, animationDefaultVariants } from "@/common/constants";
 import styles from "./ProductOrder.module.scss";
 
 type Props = {
@@ -25,19 +19,6 @@ export const ProductOrder: FC<Props> = ({ original, discount, currency, unitsMea
   const [localProductPrice, setLocalProductPrice] = useState<Omit<ProductPrice, "currency">>({ original, discount });
   const [localInputValue, setLocalInputValue] = useState(ProductsAmountOfUnitsMeasure.PCS);
   const { setNotification } = useActions();
-
-  useChangeEffect(() => {
-    const variantTitles = {
-      [ProductUnitsMeasure.BOX]: `${ProductsAmountOfUnitsMeasure.BOX} units in ${ProductUnitsMeasure.BOX}`,
-      [ProductUnitsMeasure.PACK]: `${ProductsAmountOfUnitsMeasure.PACK} units in ${ProductUnitsMeasure.PACK}`,
-    };
-
-    const title = variantTitles[currentOrderPriceVariant as keyof typeof variantTitles];
-
-    if (title) {
-      setNotification({ ...notificationInfoDefaultTemplate, title });
-    }
-  }, [currentOrderPriceVariant]);
 
   useChangeEffect(() => {
     getProductPrice({
