+++ src/components/Product/components/ProductOrder/components/ProductOrderNavigation/ProductOrderNavigation.module.scss
@@ -5,6 +5,7 @@
   display: flex;
   flex-shrink: 0;
   align-items: center;
+  position: relative;
 
   @include desktop {
     flex-direction: column;
@@ -27,56 +28,73 @@
     transition: width .3s;
   }
 
-  .productOrderInputContainer {
-    display: flex;
-    margin-right: 24px;
-    align-items: center;
-    justify-content: center;
-    border-radius: 12px;
-    border: 1px solid var(--primary-c1-color-600);
-    background-color: var(--primary-c1-color-300);
-    width: 170px;
-    padding: 1.5px 16px;
-    transition: width .3s;
-
-    @include desktop {
-      margin-right: 0;
-    }
+  .productOrderNavigationWrapper {
 
     @include largeMobile {
       width: 100%;
-      transition: width .3s;
     }
 
-    .productOrderInputWrapper {
+
+    .productOrderNotification {
+      font-size: 14px;
+      font-family: $secondary-font;
+      color: var(--secondary-c2-color-500);
+      position: absolute;
+      top: -16px;
+      left: 0;
+    }
+
+    .productOrderInputContainer {
       display: flex;
+      margin-right: 24px;
       align-items: center;
-
-      &:after {
-        content: "|";
-        margin: 0 16px;
-        width: 2px;
-        height: 100%;
-        color: var(--primary-c1-color-700);
-      }
-    }
-
-    .productOrderInput {
+      justify-content: center;
+      border-radius: 12px;
+      border: 1px solid var(--primary-c1-color-600);
       background-color: var(--primary-c1-color-300);
-      width: 25px;
+      width: 170px;
+      padding: 1.5px 16px;
       transition: width .3s;
+
+      @include desktop {
+        margin-right: 0;
+      }
 
       @include largeMobile {
         width: 100%;
         transition: width .3s;
       }
-    }
 
-    .productOrderSelect {
-      width: 100%;
-      font-size: 14px;
-      font-weight: 600;
-      transition: width .3s;
+      .productOrderInputWrapper {
+        display: flex;
+        align-items: center;
+
+        &:after {
+          content: "|";
+          margin: 0 16px;
+          width: 2px;
+          height: 100%;
+          color: var(--primary-c1-color-700);
+        }
+      }
+
+      .productOrderInput {
+        background-color: var(--primary-c1-color-300);
+        width: 25px;
+        transition: width .3s;
+
+        @include largeMobile {
+          width: 100%;
+          transition: width .3s;
+        }
+      }
+
+      .productOrderSelect {
+        width: 100%;
+        font-size: 14px;
+        font-weight: 600;
+        transition: width .3s;
+      }
     }
   }
